requirements.txt
google-genai==1.65.0
python-dotenv==1.2.2
